Reestructured PSIGHOS and PSIGHOS-RETAL
Now PSIGHOS contains the changes made to the core simulation and
PSIGHOS-RETAL the new classes

+++ PSIGHOS/src/es/ull/iis/simulation/core/TimeUnit.java
@@ -45,7 +45,7 @@ public enum TimeUnit {
 	 * @param sourceUnit Source time unit
 	 * @return The value of the source when expressed as this time unit.
 	 */
-	public long convert(long sourceValue, TimeUnit sourceUnit) {
+	public long convert(double sourceValue, TimeUnit sourceUnit) {
 		return Math.round(sourceValue * conversion[sourceUnit.ordinal()][ordinal()]);
 	}
